+++ src/e2eHelpers/dataExtractor.js
