+++ src/schema/organization/__tests__/organization.spec.js
@@ -1,8 +1,8 @@
 import { createTestClient } from 'apollo-server-testing';
 import * as queries from './api';
 import expectedResultOrganizationData from './expectedResults';
-import testServer from '../../../testUtils/integration/serverFactory';
-import getPropsFromList from '../../../testUtils/integration/dataExtractor';
+import testServer from '../../../e2eHelpers/serverFactory';
+import getPropsFromList from '../../../e2eHelpers/dataExtractor';
 
 describe('Organization type tests', () => {
   const { query } = createTestClient(testServer);
