+++ src/schema/organization/members/__tests__/members.spec.js
@@ -1,8 +1,8 @@
 import { createTestClient } from 'apollo-server-testing';
 import * as queries from './api';
 import expectedResultMembers from './expectedResults';
-import testServer from '../../../../testUtils/integration/serverFactory';
-import getPropsFromList from '../../../../testUtils/integration/dataExtractor';
+import testServer from '../../../../e2eHelpers/serverFactory';
+import getPropsFromList from '../../../../e2eHelpers/dataExtractor';
 
 describe('Member type tests', () => {
   const maxNumberOfMembers = 10;
